Added console input history.
Improved cache data security.

git-svn-id: https://cvet.by:8443/svn/FOnlineSource@582 c2c6a35e-ad2b-4e1f-bb64-97b989be360b

Former-commit-id: 1f40c02a26f277d09adce1f19e3911780d8c2a84

+++ Source/SpriteManager.cpp
@@ -3212,7 +3212,8 @@ void SpriteManager::RefreshPure3dAnimationSprite( Animation3d* anim3d )
     for( size_t i = 0, j = sprData.size(); i < j; i++ )
     {
         SpriteInfo* si = sprData[ i ];
-        if( si && si->UsedForAnim3d && !si->Anim3d && si->Width == draw_width && si->Height == draw_height && si->Atlas->Type == anim3d->SprAtlasType )
+        if( si && si->UsedForAnim3d && !si->Anim3d && (uint) si->Width == draw_width &&
+            (uint) si->Height == draw_height && si->Atlas->Type == anim3d->SprAtlasType )
         {
             index = (uint) i;
             break;
